Create listener for terminal window resizes. Try thread stuff.
The listener stores the current size of the terminal window (and consequently the size of the screen) and whether or not the screen has been resized since the last check. Created a single thread for the main game loop: this is the same as not creating a thread (was just testing/trying the syntax). Next feature should be creating a thread for player input (separate from the main game loop).