Clean more stuff up. Resize should be a menu feature.
Cleaned things relating to messages that are going to be handled differently in this game (in comparison to our hero one). Resize should become a menu feature in the future when/if we implement a 'scaling' feature: fullscreen gameplay implies a map that is too big for the size of the player (boring gameplay, hard to see); changing resolutions mid game might cause the game to need to place the player inbounds again. Resolution scaling: pick ('base') smallest screen size; bigger sizes are multiples of the base resolution (scale sprites).